+++ src/schemathesis/specs/openapi/expressions/nodes.py
@@ -58,7 +58,7 @@ class Method(Node):
     """A node for `$method` expression."""
 
     def evaluate(self, context: ExpressionContext) -> str:
-        return context.case.endpoint.method
+        return context.case.endpoint.method.upper()
 
 
 @attr.s(slots=True)  # pragma: no mutate