+++ src/schemathesis/stateful.py
@@ -1,13 +1,19 @@
 import enum
 import json
-from typing import Any, Callable, Dict, Generator, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, Dict, Generator, List, Optional, Tuple, Union
 
 import attr
 import hypothesis
+from hypothesis.stateful import RuleBasedStateMachine
+from requests.structures import CaseInsensitiveDict
+from starlette.applications import Starlette
 
 from .exceptions import InvalidSchema
 from .models import Case, Endpoint
 from .utils import NOT_SET, GenericResponse
+
+if TYPE_CHECKING:
+    from .schemas import BaseSchema
 
 
 class Stateful(enum.Enum):
@@ -92,3 +98,212 @@ class Feedback:
         for data in self.stateful_tests.values():
             endpoint = data.make_endpoint()
             yield endpoint, make_test_or_exception(endpoint, test, settings, seed)
+
+
+@attr.s(slots=True)  # pragma: no mutate
+class StepResult:
+    response: GenericResponse = attr.ib()  # pragma: no mutate
+    case: Case = attr.ib()  # pragma: no mutate
+
+
+class Direction:
+    name: str
+    status_code: str
+    endpoint: Endpoint
+
+    def set_data(self, case: Case, **kwargs: Any) -> None:
+        raise NotImplementedError
+
+
+def _print_case(case: Case) -> str:
+    endpoint = f"state.schema['{case.endpoint.path}']['{case.endpoint.method.upper()}']"
+    data = [
+        f"{name}={getattr(case, name)}"
+        for name in ("path_parameters", "headers", "cookies", "query", "body", "form_data")
+        if getattr(case, name) is not None
+    ]
+    return f"{endpoint}.make_case({', '.join(data)})"
+
+
+@attr.s(slots=True, repr=False)  # pragma: no mutate
+class _DirectionWrapper:
+    """Purely to avoid modification of `Direction.__repr__`."""
+
+    direction: Direction = attr.ib()  # pragma: no mutate
+
+    def __repr__(self) -> str:
+        path = self.direction.endpoint.path
+        method = self.direction.endpoint.method.upper()
+        return f"state.schema['{path}']['{method}'].links['{self.direction.status_code}']['{self.direction.name}']"
+
+
+class APIStateMachine(RuleBasedStateMachine):
+    """The base class for state machines generated from API schemas.
+
+    Exposes additional extension points in the testing process.
+    """
+
+    bundles: Dict[str, CaseInsensitiveDict]
+    schema: "BaseSchema"
+
+    def __init__(self) -> None:
+        super().__init__()
+        self.setup()
+
+    def _pretty_print(self, value: Any) -> str:
+        if isinstance(value, Case):
+            return _print_case(value)
+        if isinstance(value, tuple) and len(value) == 2:
+            result, direction = value
+            wrapper = _DirectionWrapper(direction)
+            return super()._pretty_print((result, wrapper))
+        return super()._pretty_print(value)
+
+    def setup(self) -> None:
+        """Hook method that runs unconditionally in the beginning of each test scenario.
+
+        Does nothing by default.
+        """
+
+    def teardown(self) -> None:
+        pass
+
+    # To provide the return type in the rendered documentation
+    teardown.__doc__ = RuleBasedStateMachine.teardown.__doc__
+
+    def transform(self, result: StepResult, direction: Direction, case: Case) -> Case:
+        raise NotImplementedError
+
+    def step(self, previous: Optional[Tuple[StepResult, Direction]], case: Case) -> StepResult:
+        """A single state machine step.
+
+        :param previous: Optional result from the previous step and the direction in which this step should be done.
+        :param Case case: Generated test case data that should be sent in an API call to the tested endpoint.
+
+        Schemathesis prepares data, makes a call and validates the received response.
+        It is the most high-level point to extend the testing process. You probably don't need it in most cases.
+        """
+        if previous is not None:
+            result, direction = previous
+            case = self.transform(result, direction, case)
+        self.before_call(case)
+        response = self.call(case)
+        self.after_call(response, case)
+        self.validate_response(response, case)
+        return self.store_result(response, case)
+
+    def before_call(self, case: Case) -> None:
+        """Hook method for modifying the case data before making a request.
+
+        :param Case case: Generated test case data that should be sent in an API call to the tested endpoint.
+
+        Use it if you want to inject static data to **all requests**, for example, an
+        API token that should always be present in headers:
+
+        .. code-block:: python
+
+            class APIWorkflow(schema.as_state_machine()):
+
+                def before_call(self, case):
+                    case.headers = case.headers or {}
+                    case.headers["Authorization"] = "Bearer <TOKEN>"
+        """
+
+    def after_call(self, response: GenericResponse, case: Case) -> None:
+        """Hook method for additional actions with case or response instances.
+
+        :param response: Response from the application under test.
+        :param Case case: Generated test case data that should be sent in an API call to the tested endpoint.
+
+        For example, you can log all response statuses by using this hook:
+
+        .. code-block:: python
+
+            import logging
+
+            logger = logging.getLogger(__file__)
+            logger.setLevel(logging.INFO)
+
+            class APIWorkflow(schema.as_state_machine()):
+
+                def after_call(self, response, case):
+                    logger.info(
+                        "%s %s -> %d",
+                        case.method,
+                        case.path,
+                        response.status_code,
+                    )
+
+            # POST /users/ -> 201
+            # GET /users/{user_id} -> 200
+            # PATCH /users/{user_id} -> 200
+            # GET /users/{user_id} -> 200
+            # PATCH /users/{user_id} -> 500
+        """
+
+    def call(self, case: Case) -> GenericResponse:
+        """Make a request to an endpoint.
+
+        :param Case case: Generated test case data that should be sent in an API call to the tested endpoint.
+        :return: Response from the application under test.
+
+        Here you can pass additional arguments to :func:`Case.call <schemathesis.models.Case.call>`, which mostly
+        are proxied to :func:`requests.request`:
+
+        .. code-block:: python
+
+            class APIWorkflow(schema.as_state_machine()):
+
+                def call(self, case):
+                    return case.call(verify=False)
+
+        The above example disables the server's TLS certificate verification.
+
+        Note that WSGI/ASGI applications are detected automatically in this method.
+        """
+        method: Callable
+        if case.app is not None:
+            if isinstance(case.app, Starlette):
+                method = case.call_asgi
+            else:
+                method = case.call_wsgi
+        else:
+            method = case.call
+        return method()
+
+    def validate_response(self, response: GenericResponse, case: Case) -> None:
+        """Validate an API response.
+
+        :param response: Response from the application under test.
+        :param Case case: Generated test case data that should be sent in an API call to the tested endpoint.
+        :raises CheckFailed: If any of the supplied checks failed.
+
+        If you need to change the default checks or provide custom validation rules, you can do it here.
+
+        .. code-block:: python
+
+            def my_check(response, case):
+                # some assertions
+
+            class APIWorkflow(schema.as_state_machine()):
+
+                def validate_response(self, response, case):
+                    case.validate_response(
+                        response,
+                        checks=(my_check, )
+                    )
+
+        The state machine from the example above will execute only the ``my_check`` check instead of all
+        available checks.
+
+        Each check function should accept ``response`` as the first argument and ``case`` as the second one and raise
+        ``AssertionError`` if the check fails.
+
+        **Note** that it is preferred to pass check functions as an argument to ``case.validate_response``.
+        In this case, all checks will be executed, and you'll receive a grouped exception that contains results from
+        all provided checks rather than only the first encountered exception.
+        """
+        case.validate_response(response)
+
+    def store_result(self, response: GenericResponse, case: Case) -> StepResult:
+        return StepResult(response, case)