+++ test/specs/openapi/test_schemas.py
@@ -8,7 +8,7 @@ def test_get_endpoint_via_remote_reference(openapi_version, swagger_20, schema_u
     resolved = swagger_20.get_endpoint_by_reference(f"{schema_url}#/paths/~1users~1{{user_id}}/patch")
     assert isinstance(resolved, Endpoint)
     assert resolved.path == "/users/{user_id}"
-    assert resolved.method == "PATCH"
+    assert resolved.method.upper() == "PATCH"
     # Via common parameters for all methods
     if openapi_version.is_openapi_2:
         assert resolved.query == {